refactor : join us location field 반응형 웹디자인 삭제
