+++ app/join-us/components/Body.tsx
@@ -2,7 +2,7 @@ import BirthGenderInputBox from "./bodyItems/BirthGenderInputBox";
 import NameInput from "./bodyItems/NameInput";
 import PasswordInputBox from "./bodyItems/password/PasswordInputBox";
 import PurposeInterestContestSelectBox from "./bodyItems/selector/PurposeInterestContestSelectBox";
-import AvailabilityLocation from "./bodyItems/Location";
+import AvailabilityLocation from "./bodyItems/location/Location";
 import Selector from "./bodyItems/selector/Selector";
 import IdInputBox from "./bodyItems/id/IdInputBox";
 import BodyExplain from "./bodyItems/BodyExplain";
