+++ app/join-us/components/bodyItems/location/Location.tsx
@@ -1,11 +1,12 @@
 // coponenets
-import LocationHeader from "./location/LocationHeader";
-import AreaField from "./location/AreaField";
-import CitiesField from "./location/CitiesField";
-import DistrctField from "./location/DistrctField";
-import SelectedLocation from "./location/SelectedLocation";
+import LocationHeader from "./LocationHeader";
+import AreaField from "./AreaField";
+import CitiesField from "./CitiesField";
+import DistrctField from "./DistrctField";
+import SelectedLocation from "./SelectedLocation";
 
 export default function AvailabilityLocation() {
+  // 굳이 이 컴포넌트가 서버일 이유는 없음
   return (
     <div id="ability-day" className="w-full">
       <LocationHeader />
